+++ zeroui/src/main/java/com/zero/zero_tools/zeroui/text/TextModels.kt
@@ -26,5 +26,8 @@ public enum class Tone {
     Success,
     Error,
     Warning,
-    Inverse
+    Inverse,
+    Info,
+    Accent,
+    Disabled
 }